+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.0.1+3]
 
 * Adds InterpolationOption
+* Allows locale and interpolation options override on `t`
 
 ## [0.0.1] - TODO: Add release date.